Pygcn File
pygcn created using the original resent18 model as the base.  Using k nearest neighbors to implement the change.  Still want to test with different learning rate schedulers, see how a few different # of neighbors might affect the model for knn part, and continue to mess with the layers, but is in working state and has been semi optimized for different variables.